% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformUI.R
\name{estimateLogicleParams}
\alias{estimateLogicleParams}
\title{Estimate parameters of Logicle Transform}
\usage{
estimateLogicleParams(ff, channel)
}
\arguments{
\item{ff}{a flowFrame}

\item{channel}{string indicating the channel for which to estimate the
transformation. Intended to run only for one channel at a time.}
}
\value{
list with parameter values w, t, m and a.
}
\description{
Estimates the transformation parameters using estimateLogicle and returns the
found parameters instead of a transformation list. Returns default parameters
and a warning if estimateLogicle fails.
}
\details{
Intended to be used internally by transformUI.
}
