% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformUI.R
\name{extractTransformListParams}
\alias{extractTransformListParams}
\title{Extract transformation parameters from transform list}
\usage{
extractTransformListParams(ff, tl)
}
\arguments{
\item{ff}{flowFrame}

\item{tl}{transformList}
}
\value{
list with transformations and parameters per channel
}
\description{
Intended to be used internally by transformUI.
}
