% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformUI.R
\name{transformUI}
\alias{transformUI}
\title{User interface for flow cytometry transformation functions.}
\usage{
transformUI(dat, sampleSize = 10000, transformList = NULL)
}
\arguments{
\item{dat}{either a \code{flowFrame} or \code{flowSet}.}

\item{sampleSize}{size of the subset used to estimate the parameters and
display in the plots. Default is 10'000.}

\item{transformList}{a previously created \code{transformList}.}
}
\value{
transformation list with the channels for which a transformation has
  been set, NULL if no transformations were defined.
}
\description{
This Shiny gadget provides a user interface for the log10, Logicle and asinh
transforms in the \code{\link[flowCore]{flowCore}} package.
}
\details{
For each parameter, the user can choose between no, log10, Logicle or asinh
transformation and adjust the parameters for the Logicle and asinh functions.
When pressing the Set button, the current transformation for the selected
channel is added to the \code{\link[flowCore]{transformList}} that is
returned when the App is closed by pressing the Done button.

For the logicle transformation, the transformation parameters are preset to
the values estimated by \code{\link[flowCore]{estimateLogicle}} for the
chosen channel. After manually adjusting the parameters, the parameters can
be reset to the estimated values by pressing the Auto button.

Because of the longer calculation time, the parameters for the asinh
transformation are only estimated when the user presses the Auto button. You
can check the calculation progress in the console. Estimation is done via
\code{\link[flowVS]{estParamFlowVS}} from the \code{\link[flowVS]{flowVS}}
package.Without pressing the Auto button, fixed default values are set.

To speed up calculations and data display, the app only works with a random
subset of the provided data. The size of that subset is defined by the
parameter \code{sampleSize} that is passed to the function.

Data can be provided as a \code{\link[flowCore]{flowFrame}} or a
\code{\link[flowCore]{flowSet}}. If a \code{flowSet} is passed to the
function, all measurements in the \code{flowSet} are combined to a single
\code{flowFrame}, and the random subset is drawn from the combined data.

Via the \code{transformList} argument, the user can pass a previously created
\code{transformList} to the function to visualize, modify and extend the
list. Note that currently all transformations not supported by transformUI
will be removed.

Not that the density plot does not update live upon manipulating the sliders
but only after the user releases the slider.

The app returns a \code{transformList} that can be used used with the
\code{\link[flowCore]{transform}} function in \code{flowCore}.
}
\author{
Stephan Benke (\email{s.benke@cytometry.uzh.ch})
}
